refactor(mm-next): rename constants/component to constants/form
