+++ packages/mirror-media-next/components/login/main-form-registration.js
@@ -18,7 +18,7 @@ import { FirebaseError } from 'firebase/app'
 import EmailInput from './email-input'
 import RegistrationPasswordInput from './registration-password-input'
 import CenteredHint from '../shared/centered-hint'
-import { InputState } from '../../constants/component'
+import { InputState } from '../../constants/form'
 import PrimaryButton from '../shared/buttons/primary-button'
 import DefaultButton from '../shared/buttons/default-button'
 import { isValidEmail, isValidPassword } from '../../utils'
